Example pipeline now working up, and including, GRIDSS assembly. Just variant detection and annotation left to do.

+++ example/example_pipeline.sh
@@ -13,14 +13,14 @@ FQ2=example.R2.fq
 REFERENCE=~/reference_genomes/human/hg19.fa
 INPUT=example.input.bam
 OUTPUT=example.sv.vcf
-ASSEMBLY=${OUTPUT/.sv.vcf/.gridss.assembly.bam}
+RAW_GRIDSS_ASSEMBLY=${OUTPUT/.sv.vcf/.gridss.assembly.bam}
 GRIDSS_JAR=~/bin/gridss-1.4.5-SNAPSHOT-jar-with-dependencies.jar
 JVM_ARGS="
 	-Dsamjdk.use_async_io_read_samtools=true 
 	-Dsamjdk.use_async_io_write_samtools=true 
 	-Dsamjdk.use_async_io_write_tribble=true"
 
-	
+GRIDSS_THRESHOLD_COVERAGE=10000
 ####################
 # Environment/external tools sanity checks
 
@@ -44,11 +44,6 @@ if [[ ! -f "$REFERENCE.fai" ]] ; then
 	echo "Missing reference genome index for $REFERENCE. Creating."
 	samtools faidx $REFERENCE
 fi
-if [[ ! -f "$REFERENCE.dict" ]] ; then
-	echo "Missing reference genome dictionary for $REFERENCE. Creating."
-	exit 1 # TODO dictionary
-fi
-
 
 ####################
 # per INPUT file steps
@@ -65,50 +60,56 @@ INPUT_WORKING_PREFIX=$INPUT_WORKING_DIR/$(basename $INPUT)
 mkdir -p $INPUT_WORKING_DIR
 
 ### bwa alignment with samtools 1.6 duplicate marking
-# First step of pipeline: alignment
-bwa mem -t $(nproc) -Y $REFERENCE $FQ1 $FQ2 | \
-samtools fixmate -m -l 0 - - | \
-# Skip compression since we're in a pipe
-# RECALCULATE_SA_SUPPLEMENTARY not needed since we aligned with -Y and without -M
-# SOFTEN_HARD_CLIPS not needed since we supplied the -Y flag to bwa
-# FIX_MATE_INFORMATION not needed since we used bwa
-java $JVM_ARGS -Dsamjdk.create_index=false \
-	-cp $GRIDSS_JAR gridss.ComputeSamTags \
-	REFERENCE_SEQUENCE=$REFERENCE \
-	WORKING_DIR=$INPUT_WORKING_DIR \
-	TMP_DIR=$INPUT_WORKING_DIR \
-	I=/dev/stdin \
-	O=/dev/stdout \
-	COMPRESSION_LEVEL=0 \
-	RECALCULATE_SA_SUPPLEMENTARY=false \
-	SOFTEN_HARD_CLIPS=false \
-	FIX_MATE_INFORMATION=false \
-	TAGS=null \
-	TAGS=R2 \
-	TAGS=Q2 \
-	TAGS=MC \
-	TAGS=MQ \
-	AS=true \
-	| \
-# Although bwa reports split reads, you can identify additional split reads
-# by feeding the soft clipped reads back to bwa
-java $JVM_ARGS -Dsamjdk.create_index=false \
-	-cp $GRIDSS_JAR gridss.SoftClipsToSplitReads \
-	REFERENCE_SEQUENCE=$REFERENCE \
-	I=/dev/stdin \
-	O=/dev/stdout \
-	ALIGNER_STREAMING=true \
-	WORKER_THREADS=$(nproc) \
-	COMPRESSION_LEVEL=0 \
-	> $INPUT_WORKING_PREFIX.unsorted.bam
-# Second step of pipeline: sorting
-samtools sort -@ $(nproc) -O BAM -l 0 -T $INPUT_WORKING_DIR $INPUT_WORKING_PREFIX.unsorted.bam | \
-samtools markdup - | \
-# THRESHOLD_COVERAGE=gridss.properties/maxCoverage
-# Using non-POSIX-compatble process substitution feature of bash to fork the output stream
-tee >(java $JVM_ARGS -cp $GRIDSS_JAR gridss.analysis.CollectGridssMetrics REFERENCE_SEQUENCE=$REFERENCE I=/dev/stdin O=$INPUT_WORKING_PREFIX THRESHOLD_COVERAGE=10000) > $INPUT
-if [[ -f $INPUT ]] ; then
+if [[ ! -f $INPUT ]] ; then
+	if [[ ! -f $INPUT_WORKING_PREFIX.unsorted.bam ]] ; then
+		# First step of pipeline: alignment
+		bwa mem -t $(nproc) -Y $REFERENCE $FQ1 $FQ2 | \
+		samtools fixmate -m - - | \
+		# Skip compression since we're in a pipe
+		# RECALCULATE_SA_SUPPLEMENTARY not needed since we aligned with -Y and without -M
+		# SOFTEN_HARD_CLIPS not needed since we supplied the -Y flag to bwa
+		# FIX_MATE_INFORMATION not needed since we used bwa
+		java $JVM_ARGS -Dsamjdk.create_index=false \
+			-cp $GRIDSS_JAR gridss.ComputeSamTags \
+			REFERENCE_SEQUENCE=$REFERENCE \
+			WORKING_DIR=$INPUT_WORKING_DIR \
+			TMP_DIR=$INPUT_WORKING_DIR \
+			I=/dev/stdin \
+			O=/dev/stdout \
+			COMPRESSION_LEVEL=0 \
+			RECALCULATE_SA_SUPPLEMENTARY=false \
+			SOFTEN_HARD_CLIPS=false \
+			FIX_MATE_INFORMATION=false \
+			TAGS=null \
+			TAGS=R2 \
+			TAGS=Q2 \
+			TAGS=MC \
+			TAGS=MQ \
+			AS=true \
+			| \
+		# Although bwa reports split reads, you can identify additional split reads
+		# by feeding the soft clipped reads back to bwa
+		java $JVM_ARGS -Dsamjdk.create_index=false \
+			-cp $GRIDSS_JAR gridss.SoftClipsToSplitReads \
+			REFERENCE_SEQUENCE=$REFERENCE \
+			I=/dev/stdin \
+			O=/dev/stdout \
+			ALIGNER_STREAMING=true \
+			WORKER_THREADS=$(nproc) \
+			> $INPUT_WORKING_PREFIX.unsorted.bam
+	fi
+	# Second step of pipeline: sorting
+	samtools sort -@ $(nproc) -O BAM -l 0 -T $INPUT_WORKING_DIR $INPUT_WORKING_PREFIX.unsorted.bam | \
+	samtools markdup - - -O BAM | \
+	# THRESHOLD_COVERAGE=gridss.properties/maxCoverage
+	# Using non-POSIX-compatble process substitution feature of bash to fork the output stream
+	tee >(java $JVM_ARGS -cp $GRIDSS_JAR gridss.analysis.CollectGridssMetrics REFERENCE_SEQUENCE=$REFERENCE I=/dev/stdin O=$INPUT_WORKING_PREFIX THRESHOLD_COVERAGE=$GRIDSS_THRESHOLD_COVERAGE) > $INPUT && \
+	samtools index $INPUT && \
+	# the unsorted input file can now be deleted
 	rm $INPUT_WORKING_PREFIX.unsorted.bam
+	###
+	# Optional step: extraction of SV-supporting reads
+	# This speeds up GRIDSS and is also useful for manual inspection of putative SVs
 fi
 ####################
 # Insert other steps here
@@ -117,8 +118,15 @@ fi
 # Not only does does it significantly increase the FDR of SV calls
 # but it strips hard clips and makes the SA and MC tags inconsistent.
 
+
 ####################
 # GRIDSS SV calls
+#
+# Preprocess of input files has been completed, so
+# we just have to perform assembly, variant detection,
+# and annotation
+#
+
 # Add BLACKLIST parameter if using hg19 to speed up runtime
 OUTPUT_WORKING_DIR=$(dirname $OUTPUT)/$(basename $OUTPUT).gridss.working
 OUTPUT_WORKING_PREFIX=$INPUT_WORKING_DIR/$(basename $INPUT)
@@ -126,25 +134,50 @@ OUTPUT_WORKING_PREFIX=$INPUT_WORKING_DIR/$(basename $INPUT)
 GRIDSS_COMMON_ARGS="
 	REFERENCE_SEQUENCE=$REFERENCE
 	INPUT=$INPUT
-	READ_PAIR_CONCORDANCE_METHOD=PERCENTAGE
 	WORKER_THREADS=$(nproc)
 	"
-java $JVM_ARGS -Dsamjdk.create_index=true \
-	-cp $GRIDSS_JAR gridss.AssembleBreakends \
-	$GRIDSS_COMMON_ARGS \
-	OUTPUT=$ASSEMBLY
 
-java $JVM_ARGS -Dsamjdk.create_index=true \
-	-cp $GRIDSS_JAR gridss.IdentifyVariants \
-	$GRIDSS_COMMON_ARGS \
-	ASSEMBLY=$ASSEMBLY \
-	OUTPUT_VCF=$OUTPUT_WORKING_PREFIX.breakpoint.vcf
-
-# AnnotateVariants combines AllocateEvidence, AnnotateReferenceCoverage and AnnotateInexactHomology
-java $JVM_ARGS -Dsamjdk.create_index=true \
-	-cp $GRIDSS_JAR gridss.AnnotateVariants \
-	$GRIDSS_COMMON_ARGS \
-	ASSEMBLY=$ASSEMBLY \
-	INPUT_VCF=$OUTPUT_WORKING_PREFIX.breakpoint.vcf \
-	OUTPUT_VCF=$OUTPUT
-
+mkdir -p $RAW_GRIDSS_ASSEMBLY.gridss.working
+PROCESSED_GRIDSS_ASSEMBLY=$RAW_GRIDSS_ASSEMBLY.gridss.working/$(basename $RAW_GRIDSS_ASSEMBLY).sv.bam
+if [[ ! -f $PROCESSED_GRIDSS_ASSEMBLY ]] ; then	
+	# GRIDSS assembly
+	java $JVM_ARGS -Dsamjdk.create_index=true \
+		-cp $GRIDSS_JAR gridss.AssembleBreakends \
+		$GRIDSS_COMMON_ARGS \
+		OUTPUT=$RAW_GRIDSS_ASSEMBLY
+	# We now need to work out which breakpoint is supported by each breakend assembly
+	# We do this by aligning the soft clipped bases of breakend assembly contigs
+	# This is exactly the same process as converting soft clipped reads into split reads 
+	java $JVM_ARGS -Dsamjdk.create_index=false \
+			-cp $GRIDSS_JAR gridss.SoftClipsToSplitReads \
+			REFERENCE_SEQUENCE=$REFERENCE \
+			I=$RAW_GRIDSS_ASSEMBLY \
+			O=/dev/stdout \
+			ALIGNER_STREAMING=true \
+			WORKER_THREADS=$(nproc) | \
+	samtools sort -@ $(nproc) -O BAM -T $RAW_GRIDSS_ASSEMBLY.gridss.working - > $PROCESSED_GRIDSS_ASSEMBLY
+	java $JVM_ARGS -cp $GRIDSS_JAR gridss.analysis.CollectGridssMetrics \
+		REFERENCE_SEQUENCE=$REFERENCE \
+		I=$RAW_GRIDSS_ASSEMBLY \
+		O=$RAW_GRIDSS_ASSEMBLY.gridss.working/$(basename $RAW_GRIDSS_ASSEMBLY) \
+		THRESHOLD_COVERAGE=$GRIDSS_THRESHOLD_COVERAGE
+fi
+if [[ ! -f $OUTPUT_WORKING_PREFIX.breakpoint.vcf ]] ; then
+	# GRIDSS variant calling
+	java $JVM_ARGS -Dsamjdk.create_index=true \
+		-cp $GRIDSS_JAR gridss.IdentifyVariants \
+		$GRIDSS_COMMON_ARGS \
+		ASSEMBLY=$PROCESSED_GRIDSS_ASSEMBLY \
+		OUTPUT_VCF=$OUTPUT_WORKING_PREFIX.breakpoint.vcf
+fi
+exit
+if [[ ! -f $OUTPUT ]] ; then
+	# GRIDSS evidence allocation and annotation
+	# AnnotateVariants combines AllocateEvidence, AnnotateReferenceCoverage and AnnotateInexactHomology
+	java $JVM_ARGS -Dsamjdk.create_index=true \
+		-cp $GRIDSS_JAR gridss.AnnotateVariants \
+		$GRIDSS_COMMON_ARGS \
+		ASSEMBLY=$ASSEMBLY \
+		INPUT_VCF=$OUTPUT_WORKING_PREFIX.breakpoint.vcf \
+		OUTPUT_VCF=$OUTPUT
+fi
